"Make the forms look a bit nicer" now with tables
This reverts commit d6a3ad317de79aaaf052a2d7c6724ff54307594e.
Turns out that when the amount of things on the left side is not equal
to the amount of things on the right side, nothing on either side aligns
with eachother. Ideally, if something extra appears on the left, it
would push everything below it down and make sure it still aligns. There
is a tool that does just that, *t a b l e s*. I realized this right as
my previous pull request got merged but only now did I fix this