add replay timestamp (#47)
* Add a Profile page, where the users can see their information.

This Profile page is currently pretty useless. In the future,
we will allow users to change their username, password, email address,
language preference, or other information there. This would also be where
users could request their account be deleted.

* fix flake8

* Fix sass formatting

* move is_good and points out of replayfile table

* add replay timestamp

* make is_good and rep_points nullable

* timestamp optional for videos

* "Make the forms look a bit nicer" now with tables

This reverts commit d6a3ad317de79aaaf052a2d7c6724ff54307594e.
Turns out that when the amount of things on the left side is not equal
to the amount of things on the right side, nothing on either side aligns
with eachother. Ideally, if something extra appears on the left, it
would push everything below it down and make sure it still aligns. There
is a tool that does just that, *t a b l e s*. I realized this right as
my previous pull request got merged but only now did I fix this

* add replay timestamp

* timestamp optional for videos

Co-authored-by: Nathaniel Rook <nrook@nrook.com>
Co-authored-by: 32th-System <intensonat@gmail.com>