Add a TTL for unverified user accounts in addition to old replays.
This involved moving some logic into shared_content.py. This feels a little
weird but I think it's the best way to share methods between "users" and
"replays". TBH it's still not super obvious the whole Django apps concept
is useful, but we might as well keep using it for now.