Create an "About" page. (#86)
ZUN does not have a lot of requests for fan projects, but one of the few
he does have is "Make sure people know you aren't me!" So, we needed a
page like this before we launched, even to beta.

I'm not 100% sure about the text, but oh well. We can improve it later.