Add "admin" page to the website. (#133)
This allows modifying rows directly if you have an admin account.

We should only ever use this in exceedingly rare situations; notably, I don't
think we should ever use it for correcting replay data that someone put in
wrong. (Instead we should build a mod flow in the website itself to handle
that.) But it's useful for programming fixes and stuff like that.