Convert timezones to UTC before serializing them.
This fixes an odd issue I had when working on the "reanalyze replay"
function. Specifically, the way it was before, Django would only
convert a naive datetime to an aware datetime when the instance was
actually saved, which lead to weird issues when comparing the "old"
replay row and the "updated" one. This also made Django log annoying
warnings.

I am still not actually sure how Touhou itself handles timezones, though.
Are the timestamps saved in replay files UTC, or are they local time?