Convert timezones to UTC before serializing them. (#177)
* Convert timezones to UTC before serializing them.

This fixes an odd issue I had when working on the "reanalyze replay"
function. Specifically, the way it was before, Django would only
convert a naive datetime to an aware datetime when the instance was
actually saved, which lead to weird issues when comparing the "old"
replay row and the "updated" one. This also made Django log annoying
warnings.

I am still not actually sure how Touhou itself handles timezones, though.
Are the timestamps saved in replay files UTC, or are they local time?

* Read all replay dates and times as UTC immediately.

Python tends to convert datetimes in weird ways so it's best to set
them all with a real timezone as soon as possible.