Set up Transifex. (#382)
Transifex is a continuous localization service:

https://www.transifex.com/

Basically, it provides a website that can be used to by translators,
and that then syncs to GitHub. My experience is that directly
translating things on GitHub is difficult even for me, so I think it
is not reasonable to expect anybody trying to help out to do it.

Fortunately for me, Transifex is free for open source projects.

This also contains a small tweak to django.po. Transifex did not like
having a message with an empty msgstr for some reason.