Translate game names.
Looks like this broke when we moved the game names to a constant.
When it was a big chain of ifs and elses, just calling pgettext actually
worked, since the function was called in a request. But when the names
got assigned to a constant, we had to move over to the lazy variants.

Fixes #428