Revamp the replay details page. (#431)
* Revamp the replay details page.

There's still work to do on this page, but I think it's already an
improvement.

* Run black

* Minor edits: add labels to boxes and change "Edit replay" to "Manage replay"

* add a bit of styling. pretty much light mode only

* A few dark changes too

* Add a small "no badges" line that appears if there are no badges.

This case is rare and should become even rarer once we add more badges,
so I think it's okay that it looks kind of cheap.

* Fiddle with light mode cards.

* Get rid of "Clear" badge, per request.

This means most PC-98 replays won't have badges, so I got rid of
the quick and dirty "No badge" message, and replaced it with a message
that appears if there are no stage splits (and thus appears for all
pc98 games and some other categories).

* Change "no stage split table" message to be more generic.

Before it specifically referred to pc98-like situations where you don't
have a replay file. But this happens for spell practice too.