Reorganize game list on left navigation bar.
It's hard to read a list of 10+ games without being organized, so this
splits the list up a bit.

The actual categorization is fairly arbitrary, which is why the categories
aren't labeled on the frontend. Historically, the English community has
divided the Windows games into "Classic", "Modern1" and "Modern2"
categories, where "Modern1" starts with Mountain of Faith and "Modern2"
includes more recent engine upgrades. However, there isn't an obvious
place to start "Modern2", since the engine upgrades were added over
multiple games. As such, I chose to use the plot; MoF-TD have a clear
thematic throughline, whereas DDC is totally unrelated to TD and serves
as a new beginning.