Add a page that allows you to edit an existing replay.
The edit replay page can be used to edit the metadata of an existing replay.
I reuse the "finalize publishing replay" page for this purpose. You can't
change the actual file, but you can change everything else.

PC-98 games aren't supported, not for any particular reason but because
this PR is big enough as it is.