Add a custom field type for Django enums.
The problem that caused me to write this PR is that enums get made available
on Django forms as strings. This does work eventually; the string gets set
on the model and is handled appropriately. But it's really annoying to
work with programmatically. For example, equality checks fail in places
you'd expect them to succeed.