Add a page allowing users to edit a replay (#484)
* Add a custom field type for Django enums.

The problem that caused me to write this PR is that enums get made available
on Django forms as strings. This does work eventually; the string gets set
on the model and is handled appropriately. But it's really annoying to
work with programmatically. For example, equality checks fail in places
you'd expect them to succeed.

* Add a page that allows you to edit an existing replay.

The edit replay page can be used to edit the metadata of an existing replay.
I reuse the "finalize publishing replay" page for this purpose. You can't
change the actual file, but you can change everything else.

PC-98 games aren't supported, not for any particular reason but because
this PR is big enough as it is.

* Last-minute fixes

* Respond to code review comments

* Add test for posting an invalid replay edit.

* Oops. Fix duplicate name