Creates and joins against a view used to get the rank of a replay.
See #507. This new function is not yet used, but I figure we can replace
the existing annotate_with_rank() with it. The new approach allows for us to
more easily get ranks in situations where the required data isn't returned
in the query, which is nice. It's a bit slower, so we may need to
convert this to a "materialized view", but I think it's still acceptable
right now, especially with the new index.